Refactor: 생성/수정 컴포넌트 로직 분리
